# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Inventario y Recetas",
        "description": "Esta es la documentación de la API para gestionar inventarios y recetas.",
        "contact": {
            "email": "soporte@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        },
        {
            "url": "https://api.tu-dominio.com",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/enterprise/store": {
            "post": {
                "tags": [
                    "Enterprise"
                ],
                "summary": "Guardar una nueva empresa",
                "description": "Valida y guarda los datos de una nueva empresa.",
                "operationId": "60d697a27c7014252673a5876299a5a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Empresa ABC"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "empresa@abc.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle Principal 123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirige a la vista de creación con un mensaje de éxito"
                    },
                    "422": {
                        "description": "Datos de entrada no válidos"
                    }
                }
            }
        },
        "/enterprise/search": {
            "get": {
                "tags": [
                    "Enterprise"
                ],
                "summary": "Buscar empresas",
                "description": "Busca empresas según los filtros proporcionados (nombre, email, teléfono).",
                "operationId": "7fe611ff3e67611cc5f554d2f20c4231",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre de la empresa a buscar",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Correo electrónico de la empresa a buscar",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Teléfono de la empresa a buscar",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de empresas que coinciden con los criterios de búsqueda"
                    }
                }
            }
        },
        "/enterprise/{id}": {
            "delete": {
                "tags": [
                    "Enterprise"
                ],
                "summary": "Eliminar una empresa",
                "description": "Elimina una empresa con base en su ID.",
                "operationId": "eaa45c1eb4cd5f3bf7104c35f0f5479b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirige a la vista de creación con un mensaje de éxito"
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    }
                }
            }
        },
...
        "/inventory": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Obtener la lista de inventario",
                "operationId": "3d1aeca30d99bdbdde2328b5d7654af8",
                "responses": {
                    "200": {
                        "description": "Lista de productos del inventario"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "814873058742a2b0affbd7c1f13a857b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "amount",
                                    "type",
                                    "unit"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Manzana"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 10
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Fruta"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "Kg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente"
                    }
                }
            }
        },
        "/inventory/{id}": {
            "put": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Actualizar un producto",
                "operationId": "c1dc4af5e91a130dfae8c5f33b8fd48c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "amount",
                                    "type",
                                    "unit"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Manzana"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 20
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Fruta"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "Kg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado correctamente"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Eliminar un producto",
                "operationId": "606740e9010bbd7ad298296e5ef6cef0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente"
                    }
                }
            }
        }
    },
...
}